refactor : getInfo에서 에러를 id값과 함께 반환하도록 수정
