+++ src/validate/isEmptyString.js
@@ -1,0 +1,3 @@
+export const isEmptyString = (string) => {
+  return string === "";
+};
