+++ src/validate/isValidStringLength.js
@@ -1,0 +1,3 @@
+export const isValidStringLength = (str, { min, max }) => {
+  return str.length >= min && str.length <= max;
+};
